+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Removed Step 1 from first time wizard, added info alert to primary contact form
+
 ## [1.0.1] - 2018-12-31
 
 ### Added